test(harness): split the suite by area and pin a golden-score corpus
The suite was one 444-line file, over the 300-LOC bar the rest of the package
keeps. Split along the seams that were already there -- sequence, erd, layout,
report -- with `allFinite` moved to a shared helper instead of a third copy.

Adds the two checks that would actually have caught this session's bugs.

A golden-score corpus of five real specs lifted from evlog (score ~64 to ~96),
run under both themes, asserting a floor a few points under the measured score
plus hard zeros for unrouted edges, node overlaps and edges through nodes. Every
bug found this session was emergent rather than functional -- 302 icons rendered
with no throw, no NaN and a green suite while 61% of them sat off-centre. A unit
test per fix would have caught none of them; a score delta catches all of them,
because the score aggregates.

Plus one unit test for the icon bbox, since that invariant is invisible to a
score: an icon whose line reaches 40px left of its own `x`. Verified it fails
against the old `x .. x + width` maths.

Floors use `>=`, so tuning a penalty weight does not break the suite. Measured
negative results (mergeEdges, ELK graph wrapping, edge-label wrapping) are
deliberately NOT tested -- those are decisions, and a test would freeze a config
value against the next person who re-measures.
